refactor: move NoResourcesToDownload and NoResourcesToShare
