+++ src/components/TeacherComponents/NoResourcesToShare/NoResourcesToShare.tsx
