+++ src/components/TeacherComponents/NoResourcesToShare/index.tsx
